add transactions interface - [CORE-849]
