+++ dapps/digital-twin-detail/src/components/overview/DigitaltwinResponse.interface.ts
@@ -1,0 +1,14 @@
+export interface DigitalTwinResponse {
+    total: number;
+    results: Result[];
+  }
+
+export interface Result {
+    address: string;
+    containers: any[];
+    created: number;
+    description: string;
+    name: string;
+    owner: string;
+    updated: number;
+  }
